fix: refactor onFilterCellActions import path for improved structure and maintainability in data-grid and doc-viewer components
