+++ plugins/main/public/components/common/data-grid/data-grid-service.ts
@@ -10,11 +10,7 @@ import {
 export const MAX_ENTRIES_PER_QUERY = 10000;
 import { tDataGridColumn } from './use-data-grid';
 import { cellFilterActions } from './cell-filter-actions';
-import {
-  FILTER_OPERATOR,
-  PatternDataSourceFilterManager,
-} from '../data-source/pattern/pattern-data-source-filter-manager';
-import { isNullish } from '../util';
+import { onFilterCellActions } from './filter-cell-actions';
 
 type ParseData<T> =
   | {
@@ -191,57 +187,6 @@ export const exportSearchToCSV = async (
     // @ts-ignore
     FileSaver?.saveAs(blobData, `events-${new Date().toISOString()}.csv`);
   }
-};
-
-export const onFilterCellActions = (
-  indexPatternId: string,
-  filters: Filter[],
-  setFilters: (filters: Filter[]) => void,
-) => {
-  return (
-    field: string,
-    operation:
-      | FILTER_OPERATOR.EXISTS
-      | FILTER_OPERATOR.IS
-      | FILTER_OPERATOR.IS_NOT,
-    values?: string | string[],
-  ) => {
-    // https://github.com/opensearch-project/OpenSearch-Dashboards/blob/4e34a7a5141d089f6c341a535be5a7ba2737d965/src/plugins/data/public/query/filter_manager/lib/generate_filters.ts#L89
-    const negated = [FILTER_OPERATOR.IS_NOT].includes(operation);
-    let _operation: FILTER_OPERATOR = operation;
-    if (isNullish(values) && ![FILTER_OPERATOR.EXISTS].includes(operation)) {
-      if (negated) {
-        _operation = FILTER_OPERATOR.EXISTS;
-      } else {
-        _operation = FILTER_OPERATOR.DOES_NOT_EXISTS;
-      }
-    }
-
-    const newFilters: Filter[] = [];
-    if (isNullish(values)) {
-      newFilters.push(
-        PatternDataSourceFilterManager.createFilter(
-          _operation,
-          field,
-          values,
-          indexPatternId,
-        ),
-      );
-    } else {
-      values = Array.isArray(values) ? values : [values];
-      values.forEach(item => {
-        newFilters.push(
-          PatternDataSourceFilterManager.createFilter(
-            _operation,
-            field,
-            item,
-            indexPatternId,
-          ),
-        );
-      });
-    }
-    setFilters([...filters, ...newFilters]);
-  };
 };
 
 const mapToDataGridColumn = (
